Working on Centipede Death Sound
-added a death sound for hitting the centipede, but in order for it to
play the segment needs to either wait until the sound is over to destroy
the centipede which makes an error the next time the player destroys a
segment, or make a new sound object which can possibly cause frame
drops.
I've commented out the line that makes it wait, but I'm still working on
it.