Added a spawn centipede function to GameController
-GameController can now spawn centipedes with any size that is input
into the spawnCentipede call in spawn waves. It'll be used to make a
balanced wave spawning algorithm.
- segmentsOut is now calculated by actually checking how many segments
are out, doing it otherwise is unreliable if the player rapid-fires a
segment.

-There is a glitch with centipedes with more than 2 body segments that
causes a crash. They follow way too closely to their earlier segment and
the crash is caused by line 79 in centipede.cs when they turn. The more
segments the centipede has the more likely it'll crash on a turn and I'm
not sure how to fix it. To force it add more Centipede body segments to
the first centipede and wait for turns.