Sound and Game Balancing From Playtester Feedback
I've been playtesting the alpha amoung friends and have made a few
changes based on that
- Wave-spawning changes to make the difficulty not ramp up so quickly,
but boss worms will now have no chance of having 0 additional segments
- fixed player deathsound playing multiple times
- fixed player momentarily respawning on a Game Over
- player now flashes green instead of red when invulnerable
- Balanced sound effects to be at the same volume or quieter than the
music, as I got many complaints about how much louder the sounds were.
- Pressing the spacebar will shoot every available bullet, which seems
to be balanced as the trade-off is a lack of control that can waste
valuable time in the later waves.