Lives and Changing code
-Fixed the x & y warnings
- Player now receives an additional life every 12k points, just like
original centipede which happens to fit within the balance of the game
so far.
-added a change area function to gamecontroller. This is where we would
play the teleportation animation and change the planet's color/rotation.
Currently it only resets the atmosphere so I can do balance testing.

I'm syncing now before I have much done with it because I'm not sure how
much I can do for the rest of today.