Title screen WIP
Pushed what I have so far. Need to move the instructions down a little
bit or I can move the title up some more. Some more art for the other
text on the screen and need a worm on the title screen somewhere! Also
trying to get the exploding asteroids workings soon.